+++ machines/tom/home.nix
@@ -1,6 +1,7 @@
 # https://github.com/nix-community/home-manager
 {
   imports = [
+    ./programs/awscli
     ./programs/gh
     ./programs/wd
   ];