+++ machines/tom/programs/awscli/default.nix
@@ -1,0 +1,18 @@
+# https://github.com/aws/aws-cli/tree/v2
+{ pkgs, ... }:
+{
+  programs.awscli = {
+    enable = true;
+    credentials = {
+      "default" = {
+        "credential_process" =
+          "${pkgs.sudo}/bin/sudo ${pkgs.coreutils}/bin/cat /run/secrets/aws/credentials";
+      };
+    };
+    settings = {
+      "default" = {
+        region = "us-east-1";
+      };
+    };
+  };
+}